linux-dpdk: crypto: Change queue amount
Changed the amount of queue pairs created for
cryptodevs to the maximal possible amount,
this is a workaround, but enables testing of
the odp4vpp project.

ODP-DPDK maps cryptodev queue pairs to cores
by the odp_cpu_id (which is thread id in my
case). So if ODP-DPDK is started with cpu
mask 0x1 it should be enough to have 1
pair of cryptodev queues. Unfortunatelly
that is not the case - later additional
threads are spawned using
odph_odpthreads_create, and ODP-DPDK
crashes spectacularily trying to access
the abscent queues. (this commit only
proposes a workaround)